+++ tests/test_io.py
@@ -3,7 +3,7 @@ import numpy as np
 from copy import deepcopy as dc
 from numpy.testing import assert_allclose
 
-from emg3d import meshes, models, fields, utils, io
+from emg3d import meshes, models, fields, utils, io, surveys, simulations, maps
 
 try:
     import h5py
@@ -228,3 +228,48 @@ def test_compare_dicts_collected(capsys):
     assert " True  ::                           cc         > another" in outstr
     assert "  {2}  :: Data       > d          > bb" in outstr
     assert "  {2}  :: 2onlydict" in outstr
+
+
+def test_known_classes(tmpdir):
+
+    frequency = 1.0
+    grid = meshes.TensorMesh([[2, 2], [3, 4], [0.5, 2]], (0, 0, 0))
+    field = fields.Field(grid)
+    sfield = fields.SourceField(grid, freq=frequency)
+    model = models.Model(grid, 1)
+    pointdip = surveys.Dipole('dip', (0, 1000, -950, 0, 0))
+    survey = surveys.Survey('Test', (0, 1000, -950, 0, 0),
+                            (-0.5, 0.5, 1000, 1000, -950, -950), frequency)
+    simulation = simulations.Simulation(
+            'Test1', survey, grid, model, gridding='same')
+
+    out = {
+        'TensorMesh': grid,
+        'Model': model,
+        'Field': field,
+        'SourceField': sfield,
+        'Survey': survey,
+        'Dipole': pointdip,
+        'Simulation': simulation,
+        'MapConductivity': maps.MapConductivity(),
+        'MapLgConductivity': maps.MapLgConductivity(),
+        'MapLnConductivity': maps.MapLnConductivity(),
+        'MapResistivity': maps.MapResistivity(),
+        'MapLgResistivity': maps.MapLgResistivity(),
+        'MapLnResistivity': maps.MapLnResistivity(),
+    }
+
+    # Simple primitive test to see if it can (de)serialize all known classes.
+    def test_it(ext):
+        io.save(tmpdir+'/test.'+ext, **out)
+        inp = io.load(tmpdir+'/test.'+ext)
+        del inp['_date']
+        del inp['_version']
+        del inp['_format']
+        assert out.keys() == inp.keys()
+
+    # Run through all format.
+    test_it('npz')
+    test_it('json')
+    if h5py:
+        test_it('h5')